Braceletbash and revamped longkick
There's now a new aerial attack, bracelet bash. Kind of sluggish but it packs a punch. Performed by doing any horizontal direction and attack. As not to fight for space with longkick, you perform longkick by performing attack midair while not holding a direction on the control stick/A or D. It's also much faster now, but in turn weaker.